requirements.txt
gradio
torch
transformers
sentence-transformers
faiss-cpu
PyMuPDF
python-dotenv
accelerate
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
psycopg2-binary
pydub
requests
google-generativeai
Werkzeug==2.3.7
gradio
torch
transformers
sentence-transformers
faiss-cpu
PyMuPDF
python-dotenv
accelerate